cookies: Use Browser Cookies with 'shiny'

Browser cookies are name-value pairs that are saved in a user's browser by a website. Cookies allow websites to persist information about the user and their use of the website. Here we provide tools for working with cookies in 'shiny' apps, in part by wrapping the 'js-cookie' JavaScript library https://github.com/js-cookie/js-cookie.

See also

Author

Maintainer: Jon Harmon jonthegeek@gmail.com (ORCID) [copyright holder]

Other contributors:

Klaus Hartl (js-cookie JavaScript library) [copyright holder]

Fagner Brack (js-cookie JavaScript library) [copyright holder]